refactor: move fn to own file
